added tests for Coffee class

+++ src/main/java/models/Guitar.java
@@ -3,7 +3,7 @@ package models;
 public class Guitar {
 
     private int qty;
-    private float price;
+    private double price;
 
     private String brand; //fender, gibson, ibanez, peavey
     private int numStrings; //6-8
